it was fine until it wasn't

+++ source/text/Format.cpp
@@ -384,12 +384,12 @@ string Format::Number(double value)
 		return "???";
 	else if(std::isinf(value))
 		return value > 0. ? "infinity" : "-infinity";
-	else if(fabs(value) > 1e15ll)
+	else if(fabs(value) > 1e15)
 	{
 		// Use scientific notation for excessively large numbers.
 		ostringstream out;
 		out.precision(3);
-		out << static_cast<double>(value);
+		out << value;
 		return out.str();
 	}
